Updated gtfs-sql.sql so it matches the latest GTFS-revision
Since the last update of the original version was written against a
version of 17 Feb 2014 it was needed to add one column in two tables
(agency_email in agency and timepoint in stop_times). That way it
matches the latest revision of 03 Feb 2016 of GTFS. Besides this here
and there I adjusted the sequence or made a column (non-)nullable.